Adds tweet search suggestion feature
Solves issues:
https://github.com/fossasia/loklak_wok_android/issues/27

Changes build.gradle files for increasing targetSdk and minSdk, updating
the gradle build tool version and adding support material design,
retrofit, gson and butterknife library.

Adds swipe refreshing and navigation to SearchActivity

Adds retrolambda library which allows using lambdas.
Uses RxBinding to get text changes in EditText as streams and delays it
acordingly so that repeated requests are not sent for every change.
A SwipeRefreshLayout is used to refresh the list of suggestion.
A click on search icon in keyboard, or on the list of suggestion
navigates to SearchActivity.

Adds realm database to store suggestions

Retrofit2 callbacks are transfored to RxJava Observables using
retrofit-rxjava adapter.
Realm database is used to store list of suggestions while the fragment
is in the lifecycle stage of onStop.