Fixes #34: Replace Processing.org's applet
Uses loklak's "suggest" API to fetch querries to scrape, these queries
are sent to twitter.js (transpiled and bundled) as a raw resource, for
scraping. The scraped data is sent back to UI thread and displayed in a
RecyclerView. The scraped tweets are pushed back to loklak server using
"push" API.