Fixes #58: Animations and state restoration (#61)
Animation for TweetPostingActivity is added when it is opened and
closed. The data type of ArrayList in PostTweetMediaAdapter is changed
to string, which are the absolute path of image files selected or
captured by the user. This way it is easier to store the images in a
Bundle and retrieve it. Similarly, latitude and longitude are also
stored in Bundle.